Removed retundant code as per mentor suggestions.
