+++ project1/data.py
@@ -15,4 +15,17 @@ class User(db.Model):
         self.password = password
         self.timestamp = datetime.now()
 
+class Book(db.Model):
+    __tablename__ = "book"
+    isbn = db.Column(db.String, primary_key = True)
+    title = db.Column(db.String, nullable = False)
+    author = db.Column(db.String, nullable = False)
+    year = db.Column(db.String, nullable = False)
+
+    def __init__(self, isbn, title, author, year):
+        self.isbn = isbn
+        self.title = title
+        self.author = author
+        self.year = year
+
